+++ game/language/letters_spa.py
@@ -3,6 +3,6 @@ class Letters:
     twoPoints = {'D': 5, 'G':2}
     threePoints = {'C': 4,'B': 2, 'M': 2, 'P': 2}
     fourPoints = {"H": 2, "F": 1, "V": 1, "Y": 1}
-    fivePoints = {'CH': 1, 'Q': 1}
-    eightPoints = {'J': 1, 'LL': 1, 'Ñ': 1, 'RR': 1, 'X':1}
+    fivePoints = {'C': 1, 'H': 1, 'Q': 1}
+    eightPoints = {'J': 1, 'L': 2, 'Ñ': 1, 'R': 2, 'X':1}
     tenPoints = {'Z': 1}