+++ tests/test_player.py
@@ -20,7 +20,11 @@ class TestPlayer(unittest.TestCase):
     def test_playerDoesntHaveWord(self):
         player1 = Player()
         player1.tiles = [Tile('T', 1), Tile('C', 1), Tile('E', 1), Tile('A', 1), Tile('D', 1), Tile('O', 1), Tile('D', 1)]
-        self.assertEqual(player1.hasWord("torpe"), True)
+        self.assertEqual(player1.hasWord("torpe"), False)
+    
+    def test_playerHasNoTiles(self):
+        player1 = Player()
+        self.assertEqual(player1.hasWord("torpe"), False)
 
     def test_validate_user_has_letters(self):
         player1 = Player()